Tests should not depend on installed translations
The Hebrew wallclock reftest breaks whenever Yosef updates the Hebrew
clock format. But the the test actually depends on the *installed*
translation, which means it's not possible to release a new version of
gnome-desktop unless you have the new version already installed. jhbuild
makes this easy, but if you're not using jhbuild this can be quite
frustrating. The tests are not intended to depend on the installed
system.